refactor: share status effect types
